+++ core/src/test/java/es/urjc/etsii/grafo/io/serializers/CSVSerializerTest.java
@@ -2,7 +2,7 @@ package es.urjc.etsii.grafo.io.serializers;
 
 import es.urjc.etsii.grafo.io.serializers.csv.CSVConfig;
 import es.urjc.etsii.grafo.io.serializers.csv.CSVSerializer;
-import es.urjc.etsii.grafo.testutil.HelperFactory;
+import es.urjc.etsii.grafo.solver.services.reference.ReferenceResultProvider;
 import es.urjc.etsii.grafo.testutil.TestInstance;
 import es.urjc.etsii.grafo.testutil.TestSolution;
 import org.junit.jupiter.api.Assertions;
@@ -13,29 +13,57 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.List;
+
+import static es.urjc.etsii.grafo.testutil.TestHelperFactory.referencesGenerator;
+import static es.urjc.etsii.grafo.testutil.TestHelperFactory.solutionGenerator;
 
 public class CSVSerializerTest {
 
-    public CSVSerializer<TestSolution, TestInstance> initCSV(char separator, Path p){
+
+    public CSVSerializer<TestSolution, TestInstance> initCSV(char separator, Path p) {
+        return initCSV(separator, p, new ArrayList<>());
+    }
+
+    public CSVSerializer<TestSolution, TestInstance> initCSV(char separator, Path p, List<ReferenceResultProvider> references) {
         var config = new CSVConfig();
         config.setSeparator(separator);
         config.setFolder(p.toFile().getAbsolutePath());
-        return new CSVSerializer<>(config);
+        return new CSVSerializer<>(config, references);
     }
 
-    @Test
-    public void writeEmptyCSV(@TempDir Path temp) throws IOException {
+
+    public void writeEmptyCSVParameters(Path temp, List<ReferenceResultProvider> providers) throws IOException {
         char separator = ',';
-        var csv = initCSV(separator, temp);
+        var csv = initCSV(separator, temp, providers);
         var csvPath = temp.resolve("test.csv");
         csv.serializeResults(new ArrayList<>(), csvPath);
 
         Assertions.assertTrue(Files.exists(csvPath));
         String csvContent = Files.readString(csvPath);
         String[] cols = csvContent.split(String.valueOf(separator));
-        // CSV must contains 6 columns: intance name, algorithm name, iteration, value, t, ttb
+        // CSV must contain 6 columns: instance name, algorithm name, iteration, value, t, ttb
         Assertions.assertEquals(cols.length, 6);
+    }
+
+    @Test
+    public void writeEmptyCSV(@TempDir Path temp) throws IOException {
+        writeEmptyCSVParameters(temp, new ArrayList<>());
+    }
+
+    @Test
+    public void writeEmptyCSVWithReferences(@TempDir Path temp) throws IOException {
+        writeEmptyCSVParameters(temp, referencesGenerator(10,10));
+    }
+
+    @Test
+    public void writeEmptyCSVWithInvalidReference(@TempDir Path temp) {
+        Assertions.assertDoesNotThrow(() -> writeEmptyCSVParameters(temp, referencesGenerator(Double.NaN,Double.NaN)));
+    }
+
+    @Test
+    public void writeEmptyCSVInvalidPath() {
+        Assertions.assertThrows(RuntimeException.class, () -> writeEmptyCSVParameters(Path.of("/doesnotexist"), referencesGenerator(Double.NaN,Double.NaN)));
     }
 
     @Test
@@ -43,20 +71,38 @@ public class CSVSerializerTest {
         char separator = ';';
         var csv = initCSV(separator, temp);
         var csvPath = temp.resolve("test2.csv");
-
-        var data = Arrays.asList(
-                HelperFactory.solutionGenerated("fakeInstance", "fakeExp", "fakeAlg", 1, 2, 10, 8),
-                HelperFactory.solutionGenerated("fakeInstance2", "fakeExp2", "fakeAlg2", 2, 4, 12, 7),
-                HelperFactory.solutionGenerated("fakeInstance3", "fakeExp3", "fakeAlg3", 3, 5, 14, 6)
-        );
+        var data = solutionGenerator();
 
         csv.serializeResults(data, csvPath);
         Assertions.assertTrue(Files.exists(csvPath));
         var csvContent = Files.readAllLines(csvPath);
         Assertions.assertEquals(data.size(), csvContent.size() - 1); // CSV has an extra row, the header
-        for(var line: csvContent){
+        for (var line : csvContent) {
             String[] cols = line.split(String.valueOf(separator));
             Assertions.assertEquals(cols.length, 6);
+        }
+    }
+
+    @Test
+    public void writeCSVDataWithCustomSeparatorWithReferences(@TempDir Path temp) throws IOException {
+        char separator = ';';
+        var references = referencesGenerator(10.10,10.10);
+        var csv = initCSV(separator, temp, references);
+        var csvPath = temp.resolve("test2.csv");
+        var data = solutionGenerator();
+
+        csv.serializeResults(data, csvPath);
+        Assertions.assertTrue(Files.exists(csvPath));
+        var csvContent = Files.readAllLines(csvPath);
+        Assertions.assertEquals(data.size() * 2, csvContent.size() - 1); // CSV has an extra row, there are references
+        for (var line : csvContent) {
+            String[] cols = line.split(String.valueOf(separator));
+            Assertions.assertEquals(cols.length, 6);
+            if(cols[1].equals("TestProvider")){
+                Assertions.assertEquals(Double.parseDouble(cols[3]), 10.10);
+                Assertions.assertEquals(Long.parseLong(cols[4]), 10.10 * 1000000000L); // Seconds to nanoseconds
+
+            }
         }
     }
 }