+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v0.13
-
+- (Fix) Use reference results in all ResultSerializers instead of only when exporting to XLSX.
 
 # v0.12
 - (New) GRASP::afterGRASP: Optionally execute anything after the GRASP constructive finishes